restaurant/views.py
from typing import Dict, Any

from django.contrib.auth import get_user_model
from django.contrib.auth.decorators import login_required
from django.db.models import QuerySet
from django.http import HttpResponseRedirect, request, HttpResponse
from django.shortcuts import render, redirect, get_object_or_404
from django.urls import reverse_lazy
from django.utils.decorators import method_decorator
from django.views import generic, View
from django.contrib.auth.mixins import LoginRequiredMixin
from django.views.generic import RedirectView

from .models import Cook, Dish, DishType, Ingredient
from .forms import (
    DishForm,
    CookCreationForm,
    CookUsernameSearchForm,
    CookExperienceUpdateForm,
    DishNameSearchForm,
    DishTypeNameSearchForm,
    IngredientNameSearchForm,
    IngredientForm,
)


@method_decorator(login_required, name="dispatch")
class IndexView(View):
    template_name = "restaurant/index.html"

    def get(self, request, *args, **kwargs) -> HttpResponse:
        num_cooks = Cook.objects.count()
        num_dishes = Dish.objects.count()
        num_dish_types = DishType.objects.count()
        num_ingredients = Ingredient.objects.count()

        num_visits = request.session.get("num_visits", 0)
        request.session["num_visits"] = num_visits + 1

        context = {
            "num_cooks": num_cooks,
            "num_dishes": num_dishes,
            "num_dish_types": num_dish_types,
            "num_ingredients": num_ingredients,
            "num_visits": num_visits + 1,
        }

        return render(request, "restaurant/index.html", context=context)


class DishTypeListView(LoginRequiredMixin, generic.ListView):
    model = DishType
    context_object_name = "dish_type_list"
    template_name = "restaurant/dishtype_list.html"
    paginate_by = 5

    def get_context_data(self, *, object_list=None, **kwargs) -> Dict[str, Any]:
        context = super(DishTypeListView, self).get_context_data(**kwargs)
        name = self.request.GET.get("name", "")
        context["search_form"] = DishTypeNameSearchForm(initial={"name": name})
        return context

    def get_queryset(self) -> QuerySet:
        queryset = DishType.objects.all()
        form = DishTypeNameSearchForm(self.request.GET)
        if form.is_valid():
            return queryset.filter(name__icontains=form.cleaned_data["name"])
        return queryset


class DishTypeDetailView(LoginRequiredMixin, generic.DetailView):
    model = DishType
    fields = "__all__"
    success_url = reverse_lazy("restaurant:dish-type-detail")


class DishTypeCreateView(LoginRequiredMixin, generic.CreateView):
    model = DishType
    fields = "__all__"
    success_url = reverse_lazy("restaurant:dish-type-list")


class DishTypeUpdateView(LoginRequiredMixin, generic.UpdateView):
    model = DishType
    fields = "__all__"
    success_url = reverse_lazy("restaurant:dish-type-list")


class DishTypeDeleteView(LoginRequiredMixin, generic.DeleteView):
    model = DishType
    success_url = reverse_lazy("restaurant:dish-type-list")


class DishListView(LoginRequiredMixin, generic.ListView):
    model = Dish
    paginate_by = 5
    queryset = Dish.objects.all().select_related("dish_type")

    def get_context_data(self, *, object_list=None, **kwargs) -> Dict[str, Any]:
        context = super(DishListView, self).get_context_data(**kwargs)
        name = self.request.GET.get("name", "")
        context["search_form"] = DishNameSearchForm(initial={"name": name})
        return context

    def get_queryset(self) -> QuerySet:
        queryset = Dish.objects.select_related("dish_type")
        form = DishNameSearchForm(self.request.GET)
        if form.is_valid():
            return queryset.filter(name__contains=form.cleaned_data["name"])
        return queryset


class DishDetailView(LoginRequiredMixin, generic.DetailView):
    model = Dish

    def dish_detail_view(self, dish_id) -> HttpResponse:
        dish = Dish.objects.get(pk=dish_id)
        dish(total_cost=dish.ingredients.all().sum())
        context = {
            "dish": dish,
        }
        return render(self, "restaurant/dish_detail.html", context=context)


class DishCreateView(LoginRequiredMixin, generic.CreateView):
    model = Dish
    form_class = DishForm
    success_url = reverse_lazy("restaurant:dish-list")


class DishUpdateView(LoginRequiredMixin, generic.UpdateView):
    model = Dish
    form_class = DishForm
    success_url = reverse_lazy("restaurant:dish-list")


class DishDeleteView(LoginRequiredMixin, generic.DeleteView):
    model = Dish
    success_url = reverse_lazy("restaurant:dish-list")


class CookListView(LoginRequiredMixin, generic.ListView):
    model = Cook
    paginate_by = 5

    def get_context_data(self, *, object_list=None, **kwargs) -> Dict[str, Any]:
        context = super(CookListView, self).get_context_data(**kwargs)
        username = self.request.GET.get("username", "")
        context["search_form"] = CookUsernameSearchForm(initial={"username": username})
        return context

    def get_queryset(self) -> QuerySet:
        queryset = get_user_model().objects.all()
        form = CookUsernameSearchForm(self.request.GET)

        if form.is_valid():
            return queryset.filter(username__icontains=form.cleaned_data["username"])

        return queryset


class CookDetailView(LoginRequiredMixin, generic.DetailView):
    model = Cook


class CookCreateView(LoginRequiredMixin, generic.CreateView):
    model = Cook
    form_class = CookCreationForm


class CookExperienceUpdateView(LoginRequiredMixin, generic.UpdateView):
    model = Cook
    form_class = CookExperienceUpdateForm
    success_url = reverse_lazy("restaurant:cook-list")


class CookDeleteView(LoginRequiredMixin, generic.DeleteView):
    model = Cook
    success_url = reverse_lazy("restaurant:cook-list")


class ToggleAssignToDishView(View):
    def post(self, request, pk) -> HttpResponseRedirect:
        cook = Cook.objects.get(id=request.user.id)
        if Dish.objects.get(id=pk) in cook.dishes.all():
            cook.dishes.remove(pk)
        else:
            cook.dishes.add(pk)
        return HttpResponseRedirect(reverse_lazy("restaurant:dish-detail", args=[pk]))


class IngredientListView(LoginRequiredMixin, generic.ListView):
    model = Ingredient
    paginate_by = 5

    def get_context_data(self, *, object_list=None, **kwargs) -> dict:
        context = super(IngredientListView, self).get_context_data(**kwargs)
        name = self.request.GET.get("name", "")
        context["search_form"] = IngredientNameSearchForm(initial={"name": name})
        return context

    def get_queryset(self) -> QuerySet:
        queryset = Ingredient.objects.all()
        form = IngredientNameSearchForm(self.request.GET)

        if form.is_valid():
            return queryset.filter(name__icontains=form.cleaned_data["name"])
        return queryset


class IngredientDetailView(LoginRequiredMixin, generic.DetailView):
    model = Ingredient


class IngredientCreateView(LoginRequiredMixin, generic.CreateView):
    model = Ingredient
    form_class = IngredientForm
    success_url = reverse_lazy("restaurant:ingredient-list")


class IngredientUpdateView(LoginRequiredMixin, generic.UpdateView):
    model = Ingredient
    form_class = IngredientForm
    success_url = reverse_lazy("restaurant:ingredient-list")


class IngredientDeleteView(LoginRequiredMixin, generic.DeleteView):
    model = Ingredient
    success_url = reverse_lazy("restaurant:ingredient-list")
